refactor(agent): phase C real decomposition — extract allowlist commands from telegram/client.rs
Codex's phase C left client.rs at 2,882 LoC with five pub fns that have
nothing to do with the Telegram HTTP client — they're bot-command
handlers that write to `allowlist.toml` and JSONL history files:

  append_to_allowlist     → fs lock + TOML append
  log_allowlist_change    → JSONL append to allowlist-history
  read_undoable_allowlist_entries → read + filter for /undo
  remove_from_allowlist   → atomic TOML rewrite
  log_false_positive      → JSONL append to fp-reports

Moved them to a new `telegram/commands.rs` (217 LoC). client.rs now
2,676 LoC (-206). telegram.rs barrel re-exports them from the new
module so the public path `telegram::append_to_allowlist` is
unchanged — no caller needed to update imports.

client.rs's integration tests that called these fns directly needed an
explicit `use super::super::commands::{…}` since they were previously
same-module. Identical behavior.

Workspace: 3,511 tests passing. clippy + fmt clean.
